Organized code, improve code efficiency
I manage to shorten up the code to one line for finding the states with the highest AVB and IBV.
I also removed the SQL function for median ABV and IBU. This way we are not relying on another language.
I broke the file into two Rmarkdown files so far. Hopefully we can get the make file to call these. We can add them back together, if it does not work.